Rename files to conform with conventions
