+++ src/components/index.jsx
@@ -3,6 +3,6 @@ import Header from './Header';
 import PostList from './PostList';
 import SEO from './SEO';
 import TagsBlock from './TagsBlock';
-import CampusMap from './map';
+import CampusMap from './Map';
 
 export { BlogList, Header, PostList, SEO, TagsBlock, CampusMap };
